+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - `Widget#addBadge()`
     - `Widget#addQueryHelpers()`
 - Added `Widget#defineAction(name, callback, skipRender)` and `Widget#revokeAction(name, skipRender)` methods to specify and revoke actions (also available via `#.actions` in queries)
+- Added `callAction` and `actionHandler` jora helpers
 
 ## 1.0.0-beta.70 (10-10-2022)